Position is now a double to support decimals.

+++ FRBDK/Glue/DialogTreePlugin/Generators/RootObjectCodeGenerator.cs
@@ -76,8 +76,8 @@ namespace DialogTreePlugin.SaveClasses
 
         public class Position
         {
-            public int x { get; set; }
-            public int y { get; set; }
+            public double x { get; set; }
+            public double y { get; set; }
         }
 
     }
